pyproject.toml
[build-system]
requires = ["hatchling>=1.25,<2.0"]
build-backend = "hatchling.build"

[project]
name = "forge"
version = "0.1.0"
description = "Terminal-first AI-powered DevOps and SRE platform."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<0.120",
    "langgraph>=0.1,<0.2",
    "langchain-anthropic>=0.1,<0.2",
    "openai>=1.0,<2.0",
    "redis>=5.0,<6.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "structlog>=24.0,<25.0",
    "typer>=0.9,<1.0",
    "rich>=13.0,<14.0",
    "kubernetes>=29.0,<30.0",
    "PyGithub>=2.0,<3.0",
    "httpx>=0.27,<0.28",
    "ragas>=0.1,<0.2",
    "uvicorn>=0.30,<0.31",
    "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.10,<2.0",
    "pytest>=8.2,<9.0",
    "pytest-asyncio>=0.23,<0.24",
    "ruff>=0.5,<0.6",
]

[project.scripts]
forge = "forge.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["forge"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "forge"]
markers = [
    "unit: fast isolated unit tests",
    "integration: integration-style tests for composed components",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["forge", "tests"]
